refactor query to separate file
for parity with mutations. done b/c adding users filling in queries to the tutorial
