+++ tutorial/web/src/graphql/useTasks.js
@@ -1,6 +1,5 @@
-import { useQuery } from "@apollo/client";
-import gql from "graphql-tag";
 import useTaskMutations from "./useTaskMutations";
+import { useAllTasksInProject } from './useTaskQueries';
 
 const useTasks = (project) => {
   const { tasks, loading } = useAllTasksInProject(project);
@@ -13,29 +12,3 @@ const useTasks = (project) => {
   };
 };
 export default useTasks;
-
-function useAllTasksInProject(project) {
-  const { data, loading, error } = useQuery(
-    gql`
-      query GetAllTasksForProject($partition: String!) {
-        tasks(query: { _partition: $partition }) {
-          _id
-          name
-          status
-        }
-      }
-    `,
-    { 
-      variables: { partition: project.partition },
-      pollInterval: 300
-    }
-  );
-  if (error) {
-    throw new Error(`Failed to fetch tasks: ${error.message}`);
-  }
-
-  // If the query has finished, return the tasks from the result data
-  // Otherwise, return an empty list
-  const tasks = data?.tasks ?? [];
-  return { tasks, loading };
-}
